Filter NVD vulns before diffs
